Add a new way to encode bridges
I'm not sure if this way is going to be better or not yet. On one had,
we can no longer have situations where where a bridge is not horizontal
or vertical, or where bridges do not actually end on an island.

However, we still have situations where bridges cannot start on an island,
and unless we duplicate each bridge (island 1 having a bridge that connects
it to island 2, and island 2 having the same bridge but reverse that
connects it to island 1), it is going to make figuring out how many
bridges are currently attached to an island harder. Maybe we could
improve upon this by encoding bridges in islands themselves, instead of
as a point on the map? Would fix the first problem, but still could add
some overhead for the second problem. Need to think on this more