addBridge now returns Maybe instead of erroring out
This allows for easier brute forcing of possible bridges in any given
step in the solving. Now we can just do some magic with addBridge and
guarding for the number of bridges an island is supposed to have to get
a list of all possible bridge combinations that would make this island
valid (without crossing other bridges or changing bridges already added
to this island). I think the next step will be to generate that list,
and from what I understand the list monad will be a good choice for
that?