Try without islands having the points themselves.
This is a work in progress. The code still works to solve bridges, but I
think a lot of it needs some cleaning up (from variable names, reworking
functions, etc). As it stands right now, this is about 2x slower then
having the islands have their points directly (as well as in the isnland
map), which is probably due to a bunch of O(n * log n) lookups during
the solve loop. That said, it is still more then fast enough, and
prevents a bunch of possible errors (where you could pass an 'old'
island to a 'newer' game iteration. I'm thinking for that reason alone,
this will probably be the best way to proceed, but lets try this on a
bigger puzzle and see how it fares (plus, get the code cleaned up)