More refactoring
Define lookupRemoteIslandPoint and getRemoteIslandPoint from
getIslandPoints. This one is odd, I think the old code may be easier to
ready, but with this we don't have to have any knowledge that we are
dealing with an infinite list and can better compose function / reduce
code reuse. I'm unsure on which of these approachs would be better, but
for now I'm going to go with this way.